file structure refactoring for better better maintenance 2nd stage
