+++ src/components/atoms/PriceTable/index.js
