+++ src/components/organisms/Map/mapStyles.js
